+++ cnccode/controller_classes/CTCustomer.inc.php
@@ -19,7 +19,7 @@ require_once($cfg['path_bu'] . '/BUSector.inc.php');
 require_once($cfg['path_dbe'] . '/DBEJOrdhead.inc.php');
 require_once($cfg['path_bu'] . '/BUPortalCustomerDocument.inc.php');
 require_once($cfg["path_bu"] . "/BURenBroadband.inc.php");
-require_once($cfg["path_bu"] . "/BURenContract.inc.php");
+require_once($cfg["path_bu"] . "/Burencontract.php");
 require_once($cfg["path_bu"] . "/BURenQuotation.inc.php");
 require_once($cfg["path_bu"] . "/BURenDomain.inc.php");
 require_once($cfg["path_bu"] . "/BURenHosting.inc.php");