+++ cnccode/controller_classes/CTRenewalsDashboard.php
@@ -1,9 +1,12 @@
 <?php
 global $cfg;
+
+use CNCLTD\Business\BURenContract;
+
 require_once($cfg["path_dbe"] . "/DBConnect.php");
 require_once($cfg['path_ct'] . '/CTCNC.inc.php');
 require_once($cfg['path_bu'] . '/BURenQuotation.inc.php');
-require_once($cfg['path_bu'] . '/BURenContract.inc.php');
+require_once($cfg['path_bu'] . '/Burencontract.php');
 require_once($cfg['path_bu'] . '/BURenBroadband.inc.php');
 require_once($cfg['path_bu'] . '/BURenDomain.inc.php');
 require_once($cfg['path_bu'] . '/BURenHosting.inc.php');