+++ cnccode/customClasses/AssetListExport/AssetListExporter.php
@@ -5,6 +5,7 @@ namespace CNCLTD\AssetListExport;
 use BUCustomer;
 use BUHeader;
 use BUPassword;
+use CNCLTD\Business\BURenContract;
 use DataSet;
 use DateInterval;
 use DateTime;
@@ -423,10 +424,6 @@ class AssetListExporter
         if ($lastContactColor) {
             $columnCoordinate = Coordinate::stringFromColumnIndex(self::LAST_CONTACT_COLUMN_INDEX + $isSummary);
             $this->setCellSolidColor($sheet, "{$columnCoordinate}{$currentRow}", $lastContactColor);
-            $operatingSystem = $tabularData->getOperatingSystem($i);
-            if (strpos(strtolower($operatingSystem), "microsoft") > -1) {
-                $this->patchManagementEligibleComputers++;
-            }
         }
     }
 
@@ -509,8 +506,10 @@ class AssetListExporter
         $updateCustomer->getRow($customerId);
         $updateCustomer->setValue(DBECustomer::noOfPCs, $tabularData->getNumberOfPcs());
         $updateCustomer->setValue(DBECustomer::noOfServers, $tabularData->getNumberOfServers());
-        $updateCustomer->setValue(DBECustomer::eligiblePatchManagement, $this->patchManagementEligibleComputers);
+        $updateCustomer->setValue(DBECustomer::eligiblePatchManagement, $tabularData->patchManagementEligibleComputers());
         $updateCustomer->updateRow();
+        $buContract = new BURenContract($this);
+        $buContract->updatePatchManagementContractForCustomer($customerId, $tabularData->patchManagementEligibleComputers());
     }
 
     /**