+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project changes will be shown here.
 ## [unreleased] 
 ### Added
 - Issue 1355 - Added a one off script to create Microsoft 365 Support items in contracts for customers.
+- Issue 1223 - Update Patch Management Contract numbers automatically
 ### Fixed
 - Issue 1360 - Creating sales orders for customers with incomplete details.